+++ PizzaStore.Domain/Models/Topping.cs
@@ -2,10 +2,10 @@ namespace PizzaStore.Domain.Models
 {
     public class Topping
     {
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string Name { get; set; }
 
-        public Topping(string name, double price)
+        public Topping(string name, decimal price)
         {
             Name = name;
             Price = price;
